ai-service/requirements.txt
# ai-service/requirement.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
openai-whisper
pydub
ffmpeg-python
python-multipart
audioop-lts  